Create /coverage to give us headline stats for this project. Refactor API code out of urls.py - which was always a stupid place for it to live.
