+++ polling_stations/templates/coverage.html
@@ -1,0 +1,16 @@
+{% extends 'base.html' %}
+{% block content %}
+  <div class="page-header">
+    <h1>Polling Station Coverage</h1>
+  </div>
+  <main>
+    <h2>
+      {{ num_district_councils }} / {{ num_councils }} ({{ perc_districts }}) with Polling Districts
+    </h2>
+    <h2>
+      {{ num_station_councils }} / {{ num_councils }} ({{ perc_stations }}) with Polling Stations
+    </h2>
+  </main>
+{% endblock %}
+
+
